Management: rename the Worker class to NewProcess
