+++ WindowsInstallerLib/src/Management/ProcessManager/NewProcess/WorkerFlags.cs
@@ -1,6 +1,6 @@
 ﻿namespace WindowsInstallerLib.Management.ProcessManager
 {
-    public static partial class Worker
+    public static partial class NewProcess
     {
         public static int ExitCode { get; set; }
     }
